re-factor ethash tests
